refactor: new subdir for adapter formatter
