+++ src/server/plugins/engine/outputFormatters/index.ts
@@ -6,10 +6,10 @@ import {
 import { type checkFormStatus } from '~/src/server/plugins/engine/helpers.js'
 import { type FormModel } from '~/src/server/plugins/engine/models/index.js'
 import { type DetailItem } from '~/src/server/plugins/engine/models/types.js'
+import { format as formatAdapterV1 } from '~/src/server/plugins/engine/outputFormatters/adapter/v1.js'
 import { format as formatHumanV1 } from '~/src/server/plugins/engine/outputFormatters/human/v1.js'
 import { format as formatMachineV1 } from '~/src/server/plugins/engine/outputFormatters/machine/v1.js'
 import { format as formatMachineV2 } from '~/src/server/plugins/engine/outputFormatters/machine/v2.js'
-import { format as formatMachineV3 } from '~/src/server/plugins/engine/outputFormatters/machine/v3.js'
 import { type FormContext } from '~/src/server/plugins/engine/types.js'
 
 type Formatter = (
@@ -23,7 +23,7 @@ type Formatter = (
 
 const formatters: Record<
   string,
-  Record<string, Formatter | typeof formatMachineV3 | undefined> | undefined
+  Record<string, Formatter | typeof formatAdapterV1 | undefined> | undefined
 > = {
   human: {
     '1': formatHumanV1
@@ -33,7 +33,7 @@ const formatters: Record<
     '2': formatMachineV2
   },
   adapter: {
-    '1': formatMachineV3
+    '1': formatAdapterV1
   }
 }
